use lock_guard, test=develop
